+++ .gitignore
@@ -40,4 +40,5 @@ nb-configuration.xml
 .env
 
 /.terraform/
+src/test/postman/reports
