+++ pom.xml
@@ -2,7 +2,7 @@
 <project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>it.gov.pagopa.swclient.mil</groupId>
+	<groupId>it.pagopa.swclient.mil</groupId>
 	<artifactId>feecalculator</artifactId>
 	<version>1.1.2</version>
 	<description>Fee Calculator Microservice for Multi-channel Integration Layer of SW Client Project</description>
@@ -12,15 +12,15 @@
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<compiler-plugin.version>3.8.1</compiler-plugin.version>
-		<surefire-plugin.version>3.0.0-M7</surefire-plugin.version>
+		<surefire-plugin.version>3.0.0-M9</surefire-plugin.version>
 		<sonar-plugin.version>3.9.1.2184</sonar-plugin.version>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
 		<quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
-		<quarkus.platform.version>2.16.0.Final</quarkus.platform.version>
+		<quarkus.platform.version>3.0.2.Final</quarkus.platform.version>
 		<skipITs>true</skipITs>
-		<common.version>1.1.5</common.version>
+		<common.version>2.0.1</common.version>
 		
 		<sonar.host.url>https://sonarcloud.io:443/</sonar.host.url>
 		<sonar.organization>pagopa</sonar.organization>
@@ -42,7 +42,7 @@
 	</dependencyManagement>
 	<dependencies>
 		<dependency>
-			<groupId>it.gov.pagopa.swclient.mil</groupId>
+			<groupId>it.pagopa.swclient.mil</groupId>
 			<artifactId>common</artifactId>
 			<version>${common.version}</version>
 		</dependency>
@@ -103,6 +103,10 @@
 			<artifactId>testcontainers</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+		    <groupId>io.quarkus</groupId>
+		    <artifactId>quarkus-rest-client-reactive-jackson</artifactId>
+		    </dependency>
 	</dependencies>
 	<build>
 		<plugins>
