--- src/main/java/it/gov/pagopa/swclient/mil/feecalculator/bean/PaymentMethods.java
@@ -1,10 +1,0 @@
-package it.gov.pagopa.swclient.mil.feecalculator.bean;
-
-public enum PaymentMethods {
-	PAGOBANCOMAT,
-    DEBIT_CARD,
-    CREDIT_CARD,
-    PAYMENT_CARD,
-    BANK_ACCOUNT,
-    CASH
-}
