+++ src/main/java/it/pagopa/swclient/mil/feecalculator/bean/PaymentMethods.java
@@ -1,0 +1,10 @@
+package it.pagopa.swclient.mil.feecalculator.bean;
+
+public enum PaymentMethods {
+	PAGOBANCOMAT,
+    DEBIT_CARD,
+    CREDIT_CARD,
+    PAYMENT_CARD,
+    BANK_ACCOUNT,
+    CASH
+}
