+++ src/main/resources/application.properties
@@ -10,16 +10,16 @@ quarkus.log.console.format=%d{yyyy-MM-dd HH:mm:ss.SSS} [%X{requestId}] [%p] [%c{
 
 quarkus.log.min-level=DEBUG
 #%dev.quarkus.log.level=DEBUG
-%dev.quarkus.log.category."it.gov.pagopa.swclient.mil.feecalculator".level=DEBUG
+%dev.quarkus.log.category."it.pagopa.swclient.mil.feecalculator".level=DEBUG
 
 %playground.quarkus.log.level=ERROR
-%playground.quarkus.log.category."it.gov.pagopa.swclient.mil.feecalculator".level=DEBUG
+%playground.quarkus.log.category."it.pagopa.swclient.mil.feecalculator".level=DEBUG
 
 %test.quarkus.log.level=ERROR
-%test.quarkus.log.category."it.gov.pagopa.swclient.mil.feecalculator".level=DEBUG
+%test.quarkus.log.category."it.pagopa.swclient.mil.feecalculator".level=DEBUG
 
 %prod.quarkus.log.level=${feecalculator.quarkus-log-level}
-%prod.quarkus.log.category."it.gov.pagopa.swclient.mil.feecalculator".level=${feecalculator.app-log-level}
+%prod.quarkus.log.category."it.pagopa.swclient.mil.feecalculator".level=${feecalculator.app-log-level}
 
 # ------------------------------------------------------------------------------
 # Quarkus reactive rest client configuration - GEC
@@ -42,6 +42,7 @@ fees.max.occurrences=10
 # ------------------------------------------------------------------------------
 
 %dev.quarkus.rest-client.mil-rest-api.url=https://mil-d-apim.azure-api.net
+#%dev.quarkus.rest-client.mil-rest-api.url=http://localhost:9999
 %test.quarkus.rest-client.mil-rest-api.url=https://mil-d-apim.azure-api.net
 %prod.quarkus.rest-client.mil-rest-api.url=${mil.rest-service.url}
 
@@ -73,3 +74,9 @@ gec.touchpoint.map.POS=POS
 gec.touchpoint.map.TOTEM=ANY
 gec.touchpoint.map.CASH_REGISTER=ANY
 gec.touchpoint.map.CSA=ANY
+
+# log request response of Rest Client request
+quarkus.rest-client.logging.scope=request-response
+quarkus.rest-client.logging.body-limit=1024
+quarkus.log.category."org.jboss.resteasy.reactive.client.logging".level=DEBUG
+# log request response of Rest Client request
